feat: Add Line schema with a list of Spans
This commit introduces a new `Line` schema to represent a line of text within a PDF document.  This schema contains a list of `Span` objects, which represent individual text spans within the line.  This new schema is added to the existing structure to allow for better representation of the content of the pdf.