feat: Add Page schema for representing PDF page content
Adds a `Page` schema to represent the content of a PDF page.  This includes a list of `Line` objects and a list of `fitz.Widget` objects (annotations) extracted from the page.

The schema includes:
- `page_number`: Optional integer representing the page number.
- `lines`: List of `Line` objects representing the text content.
- `widgets`: Optional list of `fitz.Widget` objects representing annotations.

Also, enables `arbitrary_types_allowed` in the Config to allow `fitz.Widget` objects in the schema.