feat: Add abstract property text_from_document
This commit introduces an abstract property `text_from_document` to the `BaseParserPDF` class. This property is intended to be implemented by subclasses to provide access to the text content of the PDF document.

The addition of this abstract property enforces a standard way for parsers to expose the textual data, promoting consistency and extensibility across different PDF parsing implementations.