feat: Implement add_widgets_to_lines_on_page in TextProcessor
This commit implements the `add_widgets_to_lines_on_page` method in the `TextProcessor` class. This method iterates through each line on the page and checks if any widget is located on the same line using `_geometry_utils.is_same_line`. If a widget is on the same line as a line object, the widget is added to the `line.widgets` list.