feat: Add abstract method `is_word_in_column`
This commit introduces a new abstract method `is_word_in_column` to the `GeometryBaseUtils` class. This method is intended to provide a standardized way to determine if a word rectangle falls within a given column rectangle. This will be crucial for more robust and accurate text extraction and layout analysis in subsequent implementations.