feat: Add CPU usage monitoring utility
This commit introduces a new utility for monitoring CPU usage during the execution of functions. It includes a decorator (`cpu_monitor_decorator`) that can be used to wrap functions and collect CPU usage data using `psutil`.

The collected data can be either stored in a global `cpu_usage_results` dictionary or saved to a Redis instance, depending on the configuration provided to the decorator. This functionality allows for easier performance profiling and debugging of CPU-intensive tasks.