refactor: Rename rect1 and rect2 to first_rect and second_rect
This commit refactors the `is_same_line` method in both the `GeometryBaseUtils` and `GeometryUtils` classes to improve readability. The parameters `rect1` and `rect2` were renamed to `first_rect` and `second_rect` respectively, making the code more explicit and easier to understand. There are no breaking changes as a result of this renaming.