feat: Refactor widget data to separate Text and Table widgets
The widget data is refactored to separate text and table widgets into separate dictionaries under the "Text" and "Table" keys respectively. This allows for easier access and management of different types of widgets. The text widgets are now filtered to exclude widgets that are already identified as part of a table.