refactor: Allow default values for parser and validator
This change adds default values for the `widget_parser_type` and `validator_type` parameters in the `main_check_pdf_fields` function.
This makes the function easier to use in cases where the default parser and validator are sufficient. This commit improves the flexibility of the function by allowing users to omit these parameters when using the default implementations.